requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
PyMySQL==1.1.0
python-dotenv==1.0.1
Werkzeug==3.1.2
reportlab==4.2.5
weasyprint==62.3
pandas==2.2.3
